@@ -1,5 +1,6 @@
 // @flow
 import React from 'react'
+import PropTypes from 'prop-types'
 
 export type RenderProps = {
   render?: any,
@@ -7,24 +8,29 @@ export type RenderProps = {
 }
 
 type RenderComponentProps = {
-  children: string
+  value: string | Array<any>
 } & RenderProps
 
-const Render = ({ render, className, children }: RenderComponentProps) => {
-  if (render) {
+export default class Render extends React.Component<RenderComponentProps> {
+  props: RenderComponentProps
+
+  static contextTypes = {
+    linguiDefaultRender: PropTypes.any
+  }
+
+  render () {
+    const { className, value } = this.props
+    const render = this.props.render || this.context.linguiDefaultRender || 'span'
+
     // Built-in element: h1, p
     if (typeof render === 'string') {
-      return React.createElement(render, {}, children)
+      return React.createElement(render, { className }, value)
     }
 
     return React.isValidElement(render)
       // Custom element: <p className="lear' />
-      ? React.cloneElement(render, {}, children)
+      ? React.cloneElement(render, {}, value)
       // Custom component: ({ translation }) => <a title={translation}>x</a>
-      : React.createElement(render, { translation: children })
+      : React.createElement(render, { translation: value })
   }
-
-  return <span className={className}>{children}</span>
 }
-
-export default Render
