@@ -1,7 +1,7 @@
 // @flow
 import React from 'react'
-import WithI18n from './WithI18n'
-import type { WithI18nProps } from './WithI18n'
+import withI18n from './withI18n'
+import type { withI18nProps } from './withI18n'
 
 import type { RenderProps } from './Render'
 import Render from './Render'
@@ -15,19 +15,23 @@ type PluralProps = {
   few?: any,
   many?: any,
   other: any,
-} & WithI18nProps & RenderProps
+} & withI18nProps & RenderProps
 
 type SelectProps = {
   value: any,
   other: any,
-} & WithI18nProps & RenderProps
+} & withI18nProps & RenderProps
 
-const Select = WithI18n()(class Select extends React.Component<*, SelectProps, *> {
+const Select = withI18n()(class Select extends React.Component<*, SelectProps> {
   props: SelectProps
 
   render () {
     const { className, render, i18n, ...selectProps } = this.props
-    return <Render className={className} render={render}>{i18n.select(selectProps)}</Render>
+    return <Render
+      className={className}
+      render={render}
+      value={i18n.select(selectProps)}
+    />
   }
 })
 
@@ -35,7 +39,7 @@ const PluralFactory = (ordinal = false) => {
   const displayName = !ordinal ? 'Plural' : 'SelectOrdinal'
   const pluralType = !ordinal ? 'plural' : 'selectOrdinal'
 
-  return class extends React.Component<*, PluralProps, *> {
+  return class extends React.Component<*, PluralProps> {
     displayName = displayName
 
     props: PluralProps
@@ -59,16 +63,18 @@ const PluralFactory = (ordinal = false) => {
       })
 
       return (
-        <Render className={className} render={render}>
-          {i18n[pluralType](pluralProps)}
-        </Render>
+        <Render
+          className={className}
+          render={render}
+          value={i18n[pluralType](pluralProps)}
+        />
       )
     }
   }
 }
 
-const Plural = WithI18n()(PluralFactory(false))
-const SelectOrdinal = WithI18n()(PluralFactory(true))
+const Plural = withI18n()(PluralFactory(false))
+const SelectOrdinal = withI18n()(PluralFactory(true))
 
 export { Plural, SelectOrdinal, Select }
 export type { PluralProps }
